requirements.txt
mypy==1.13.0
mypy-extensions==1.0.0
pygame-ce==2.5.2
typing_extensions==4.12.2
